fix: signinwidget lose context of idx pipeline in case of browser refresh
When browser refreshes in the middle of the pipelie, then SIW gets the new token and looses current context.
To support the url refresh, store stateToken in sessionStorage. If user closes the tab, they loose the context.

When to clean up sessionStorage
1. When login flow ends with success
2. When login flow ends at terminal state
3. When login flow ends with Cancel remediation
4. When login flow ends with user error, where only option is to refresh the page. for eg temrinal form errors.

Actual sessionStoage handled by okta-core, SIW only cleanup it up.